no more separate main.ts
